fix: db_query::own_server_data
