+++ commands/src/db_query.rs
@@ -4,7 +4,7 @@ use sled::Db;
 
 pub(crate) mod other_server_data;
 pub(crate) mod other_server_times_data;
-// pub(crate) mod own_server_data;
+pub(crate) mod own_server_data;
 // pub(crate) mod own_server_times_data;
 
 pub trait SledTable {
