+++ commands/src/db_query/own_server_data.rs
@@ -2,92 +2,101 @@ use super::*;
 use crate::own_server::OwnServerData;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
+use anyhow::Result;
+use sled::Db;
 
-/// 自身のマスターwebhookを a_server_data テーブルにupsertする
-pub(crate) async fn upsert_own_server_data(
-    connection: &SqlitePool,
-    server_data: &OwnServerData,
-    // server_name: &str,
-    // guild_id: &str,
-    // master_channel_id: &str,
-    // master_webhook_url: &str,
-    // private_key_pem: &str,
-    // public_key_pem: &str,
-) -> anyhow::Result<()> {
-    let guild_id = server_data.guild_id.to_string();
-    let master_channel_id = server_data.master_channel_id.to_string();
-    sqlx::query!(
-        r#"
-        INSERT INTO a_server_data (server_name, guild_id, master_channel_id, master_webhook_url, private_key_pem, public_key_pem)
-        VALUES(?, ?, ?, ?, ?, ?)
-        ON CONFLICT(guild_id) DO UPDATE SET server_name = ?, master_channel_id = ?, master_webhook_url = ?, private_key_pem = ?, public_key_pem = ?;
-        "#,
-        server_data.server_name,
-        guild_id,
-        master_channel_id,
-        server_data.master_webhook_url,
-        server_data.private_key_pem,
-        server_data.public_key_pem,
-        server_data.server_name,
-        master_channel_id,
-        server_data.master_webhook_url,
-        server_data.private_key_pem,
-        server_data.public_key_pem,
-    )
-    .execute(connection)
-    .await?;
-
-    Ok(())
+pub struct OwnServerDataTable<'a> {
+    db: &'a sled::Db,
 }
 
-// 取得する
-pub(crate) async fn select_own_server_data(
-    connection: &SqlitePool,
-    guild_id: u64,
-) -> anyhow::Result<OwnServerData> {
-    let guild_id = guild_id.to_string();
-    let row = sqlx::query!(
-        r#"
-        SELECT * FROM a_server_data WHERE guild_id = ?;
-        "#,
-        guild_id
-    )
-    .fetch_one(connection)
-    .await?;
-
-    let a_server_data = OwnServerData::from_row(
-        &row.guild_id,
-        &row.server_name,
-        &row.master_channel_id.to_string(),
-        &row.master_webhook_url,
-        &row.private_key_pem,
-        &row.public_key_pem,
-    )?;
-
-    Ok(a_server_data)
+impl<'a> OwnServerDataTable<'a> {
+    pub fn new(db: &'a sled::Db) -> Self {
+        Self { db }
+    }
 }
 
-// 取得 guild_idを用いない
-// このデータは1つしかないことを前提にしている
-pub(crate) async fn select_own_server_data_without_guild_id(
-    connection: &SqlitePool,
-) -> anyhow::Result<OwnServerData> {
-    let row = sqlx::query!(
-        r#"
-        SELECT * FROM a_server_data;
-        "#
-    )
-    .fetch_one(connection)
-    .await?;
+impl<'a> SledTable for OwnServerDataTable<'a> {
+    const TABLE_NAME: &'static str = "OwnServerDataTable";
+    // const KEY_NAME: &'static str = "OwnServerData";
+    // 自身のサーバーデータは常に1つしかない
+    // keyにどんな値を渡したとしても同じキーを使うように
+    // それぞれにkeyの引数を減らして実装したいが，その方法がわからないのでこの手段をとった
+    // クソコードだね
 
-    let a_server_data = OwnServerData::from_row(
-        &row.guild_id,
-        &row.server_name,
-        &row.master_channel_id.to_string(),
-        &row.master_webhook_url,
-        &row.private_key_pem,
-        &row.public_key_pem,
-    )?;
+    type SledKey = String;
 
-    Ok(a_server_data)
+    type SledValue = OwnServerData;
+
+    fn get_db(&self) -> &sled::Db {
+        self.db
+    }
+
+    fn upsert(&self, _key: &Self::SledKey, value: &Self::SledValue) -> Result<()> {
+        let key = "OwnServerData";
+        let value = serde_json::to_string(value)?;
+        let byte_key = value.as_bytes();
+        let db = self.get_db();
+        db.open_tree(Self::TABLE_NAME)?.insert(key, byte_key)?;
+        Ok(())
+    }
+
+    fn read(&self, _key: &Self::SledKey) -> Result<Option<Self::SledValue>> {
+        let db = self.get_db();
+        let byte_key = "OwnServerData";
+        let ret = db.open_tree(Self::TABLE_NAME)?.get(byte_key)?;
+        match ret {
+            Some(ivec) => {
+                let string = String::from_utf8(ivec.to_vec())?;
+                let value = serde_json::from_str::<Self::SledValue>(&string)?;
+                Ok(Some(value))
+            }
+            None => Ok(None),
+        }
+    }
+
+    fn read_all(&self) -> Result<Vec<Self::SledValue>> {
+        let db = self.get_db();
+        let mut ret = Vec::new();
+        let tree = db.open_tree(Self::TABLE_NAME)?;
+        for item in tree.iter() {
+            let (_key, value) = item?;
+            let string = String::from_utf8(value.to_vec())?;
+            let value = serde_json::from_str::<Self::SledValue>(&string)?;
+            ret.push(value);
+        }
+        Ok(ret)
+    }
+
+    fn delete(&self, _key: &Self::SledKey) -> Result<()> {
+        let db = self.get_db();
+        let byte_key = "OwnServerData";
+        db.open_tree(Self::TABLE_NAME)?.remove(byte_key)?;
+        Ok(())
+    }
+}
+
+
+impl OwnServerData {
+    // 必ず1つしか存在しないデータなので，keyはなんでもいい
+    // traitの方でkeyは固定している
+    pub fn db_upsert(&self, db: &Db) -> Result<()> {
+        let key = String::from("any key becouse no use");
+        let own_server_table = OwnServerDataTable::new(db);
+        own_server_table.upsert(&key, self)?;
+        Ok(())
+    }
+
+    pub fn db_read(db: &Db) -> Result<Option<Self>> {
+        let key = String::from("any key becouse no use");
+        let own_server_table = OwnServerDataTable::new(db);
+        let ret = own_server_table.read(&key)?;
+        Ok(ret)
+    }
+
+    pub fn db_delete(db: &Db) -> Result<()> {
+        let key = String::from("any key becouse no use");
+        let own_server_table = OwnServerDataTable::new(db);
+        own_server_table.delete(&key)?;
+        Ok(())
+    }
 }
