+++ commands/src/own_server.rs
@@ -37,75 +37,75 @@ impl OwnServerData {
     }
 }
 
-pub struct OwnServerDataTable<'a> {
-    db: &'a sled::Db,
-}
+// pub struct OwnServerDataTable<'a> {
+//     db: &'a sled::Db,
+// }
 
-impl<'a> OwnServerDataTable<'a> {
-    pub fn new(db: &'a sled::Db) -> Self {
-        Self { db }
-    }
-}
+// impl<'a> OwnServerDataTable<'a> {
+//     pub fn new(db: &'a sled::Db) -> Self {
+//         Self { db }
+//     }
+// }
 
-impl<'a> SledTable for OwnServerDataTable<'a> {
-    const TABLE_NAME: &'static str = "OwnServerDataTable";
-    // const KEY_NAME: &'static str = "OwnServerData";
-    // 自身のサーバーデータは常に1つしかない
-    // keyにどんな値を渡したとしても同じキーを使うように
-    // それぞれにkeyの引数を減らして実装したいが，その方法がわからないのでこの手段をとった
-    // クソコードだね
+// impl<'a> SledTable for OwnServerDataTable<'a> {
+//     const TABLE_NAME: &'static str = "OwnServerDataTable";
+//     // const KEY_NAME: &'static str = "OwnServerData";
+//     // 自身のサーバーデータは常に1つしかない
+//     // keyにどんな値を渡したとしても同じキーを使うように
+//     // それぞれにkeyの引数を減らして実装したいが，その方法がわからないのでこの手段をとった
+//     // クソコードだね
 
-    type SledKey = String;
+//     type SledKey = String;
 
-    type SledValue = OwnServerData;
+//     type SledValue = OwnServerData;
 
-    fn get_db(&self) -> &sled::Db {
-        self.db
-    }
+//     fn get_db(&self) -> &sled::Db {
+//         self.db
+//     }
 
-    fn upsert(&self, _key: &Self::SledKey, value: &Self::SledValue) -> Result<()> {
-        let key = "OwnServerData";
-        let value = serde_json::to_string(value)?;
-        let byte_key = value.as_bytes();
-        let db = self.get_db();
-        db.open_tree(Self::TABLE_NAME)?.insert(key, byte_key)?;
-        Ok(())
-    }
+//     fn upsert(&self, _key: &Self::SledKey, value: &Self::SledValue) -> Result<()> {
+//         let key = "OwnServerData";
+//         let value = serde_json::to_string(value)?;
+//         let byte_key = value.as_bytes();
+//         let db = self.get_db();
+//         db.open_tree(Self::TABLE_NAME)?.insert(key, byte_key)?;
+//         Ok(())
+//     }
 
-    fn read(&self, _key: &Self::SledKey) -> Result<Option<Self::SledValue>> {
-        let db = self.get_db();
-        let byte_key = "OwnServerData";
-        let ret = db.open_tree(Self::TABLE_NAME)?.get(byte_key)?;
-        match ret {
-            Some(ivec) => {
-                let string = String::from_utf8(ivec.to_vec())?;
-                let value = serde_json::from_str::<Self::SledValue>(&string)?;
-                Ok(Some(value))
-            }
-            None => Ok(None),
-        }
-    }
+//     fn read(&self, _key: &Self::SledKey) -> Result<Option<Self::SledValue>> {
+//         let db = self.get_db();
+//         let byte_key = "OwnServerData";
+//         let ret = db.open_tree(Self::TABLE_NAME)?.get(byte_key)?;
+//         match ret {
+//             Some(ivec) => {
+//                 let string = String::from_utf8(ivec.to_vec())?;
+//                 let value = serde_json::from_str::<Self::SledValue>(&string)?;
+//                 Ok(Some(value))
+//             }
+//             None => Ok(None),
+//         }
+//     }
 
-    fn read_all(&self) -> Result<Vec<Self::SledValue>> {
-        let db = self.get_db();
-        let mut ret = Vec::new();
-        let tree = db.open_tree(Self::TABLE_NAME)?;
-        for item in tree.iter() {
-            let (_key, value) = item?;
-            let string = String::from_utf8(value.to_vec())?;
-            let value = serde_json::from_str::<Self::SledValue>(&string)?;
-            ret.push(value);
-        }
-        Ok(ret)
-    }
+//     fn read_all(&self) -> Result<Vec<Self::SledValue>> {
+//         let db = self.get_db();
+//         let mut ret = Vec::new();
+//         let tree = db.open_tree(Self::TABLE_NAME)?;
+//         for item in tree.iter() {
+//             let (_key, value) = item?;
+//             let string = String::from_utf8(value.to_vec())?;
+//             let value = serde_json::from_str::<Self::SledValue>(&string)?;
+//             ret.push(value);
+//         }
+//         Ok(ret)
+//     }
 
-    fn delete(&self, _key: &Self::SledKey) -> Result<()> {
-        let db = self.get_db();
-        let byte_key = "OwnServerData";
-        db.open_tree(Self::TABLE_NAME)?.remove(byte_key)?;
-        Ok(())
-    }
-}
+//     fn delete(&self, _key: &Self::SledKey) -> Result<()> {
+//         let db = self.get_db();
+//         let byte_key = "OwnServerData";
+//         db.open_tree(Self::TABLE_NAME)?.remove(byte_key)?;
+//         Ok(())
+//     }
+// }
 
 #[derive(Serialize, Deserialize, Debug, Eq, PartialEq, FromRow)]
 pub struct OwnTimesData {
